Remove render loop
