+++ Cargo.toml
@@ -32,7 +32,6 @@ embedded-graphics-simulator = [
 crossterm = { version = ">=0.28.0, <0.30.0", optional = true }
 embedded-graphics-core = { version = "^0.4", optional = true }
 embedded-graphics = { version = "^0.8", optional = true }
-embassy-futures = "^0.1"
 u8g2-fonts = { version = "0.7.2", optional = true, features = [
   "embedded_graphics_textstyle",
 ] }
@@ -47,7 +46,6 @@ embedded-graphics-simulator = { version = "0.7.0", default-features = false, fea
 
 [dev-dependencies]
 tinytga = "^0.5"
-smol = "2.0.2"
 
 [profile.release]
 debug = true
