+++ src/lib.rs
@@ -12,7 +12,6 @@ pub mod image;
 pub mod layout;
 pub mod primitives;
 pub mod render;
-pub mod render_loop;
 pub mod render_target;
 pub mod surface;
 pub mod transition;
